% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff-get.R
\name{get_ff}
\alias{get_ff}
\title{Get fantasy football data}
\usage{
get_ff(url = NULL, resp = NULL, ...)
}
\arguments{
\item{url, resp}{Functionally, equivalent to the same arguments `get_scores()`.}

\item{...}{dots. Passed to wrapped functions.}
}
\value{
tibble, with the following columns: TBD.
}
\description{
Get fantasy football data.
}
\details{
Like `get_scores()`, this functions calls
`make_url*()`, followed by `request*()`, followed by `parse*()`.
}
