% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{get_scores}
\alias{get_scores}
\title{Get scores}
\usage{
get_scores(league = c("nfl", "nba"), url = NULL, ..., as = c("parsed",
  "raw"), depth = c("all", "custom"), col_names = NULL)
}
\arguments{
\item{league}{character. Name of league for which to get scores. One of `nfl` or `nba`.
Only `nfl` is currently implemented.}

\item{url}{character. Not required. If provided,
it is used with priority over `...`. Passed to `request_scores()`.
Expects `"events"` element at the top level of the response.}

\item{...}{dots. Arguments to pass to `make_url_scores()` to create `url`.}

\item{as}{character. desired type of output. One of `parsed` or `raw`.
Inspired somewhat by the `httr::content()` analogue, although the implementation is different.}

\item{depth}{character. Specifies which elements to return from parsed content.
Only specific values are allowed. Default is provided.}

\item{col_names}{character vector. Names of columns in returned `tibble`. Only used
if `depth` = `"custom"`.}
}
\value{
`tibble`
}
\description{
Get scores.
}
\details{
This functions calls `make_url_scores()` and `rqeuest_scores()`,
then provides the added functionality of parsing.
}
