% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{get_scores_nfl}
\alias{get_scores_nfl}
\title{Get NFL scores}
\usage{
get_scores_nfl(league = "nfl", col_names = c("eventId", "eventDate",
  "venueFullName", "venueCity", "venueState", "teamNameHome",
  "teamAbbreviationHome", "teamScoreHome", "teamNameAway",
  "teamAbbreviationAway", "teamScoreAway"), ...)
}
\arguments{
\item{league}{character(s). Required. Default is provided.}

\item{col_names}{character vector. Required. Default is provided.}

\item{...}{dots. Passed to wrapped function.}
}
\value{
character representing a url.
}
\description{
Get NFL scores.
}
\details{
This is a wrapper function for `get_scores()`.
}
