% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{make_url_scores_nfl}
\alias{make_url_scores_nfl}
\title{Build a url for NFL scores}
\usage{
make_url_scores_nfl(league = "nfl", seasontype = 2, week = 1,
  year = format(Sys.Date() - 365, "\%Y"), ...)
}
\arguments{
\item{league}{character. `"nfl"`.}

\item{seasontype, week}{numberic(s). Required. Defaults are provided.}

\item{year}{character and/or numeric. Default is providied.}

\item{...}{dots. Passed to wrapped function.}
}
\value{
character representing a url.
}
\description{
Build a url for NFL scores.
}
\details{
This is a wrapper function for `make_url_scores()`.
}
