% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff-parse.R
\name{parse_ff}
\alias{parse_ff}
\title{Parse fantasy football data}
\usage{
parse_ff(resp, as = c("complete", "minimal"), ...)
}
\arguments{
\item{resp}{response object (from `{httr}` package). Required.}

\item{as}{character. Desired type of output. One of `complete` or `minimal`.
Inspired somewhat by the `httr::content()` analogue, although the implementation is different.}

\item{...}{dots. Passed to `httr::content()`.}
}
\value{
tibble
}
\description{
Parse fantasy football data.
}
\details{
None.
}
\seealso{
\url{https://dusty-turner.netlify.com/post/mathlete-fantasy-football-analysis/}.
}
