% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{request_scores}
\alias{request_scores}
\title{Retrieve scores}
\usage{
request_scores(url = NULL, league = c("nfl", "nba"), ...)
}
\arguments{
\item{url}{character. Not required. If provided,
it is used with priority over `...`. Passed to `request_scores()`.
Expects `"events"` element at the top level of the response.}

\item{league}{character. Name of league for which to get scores. One of `nfl` or `nba`.
Only `nfl` is currently implemented.}

\item{...}{dots. Arguments to pass to `make_url_scores()` to create `url`.}
}
\value{
`tibble`
}
\description{
Retrieve scores.
}
\details{
Note that this functions does not come with suffixed league forms
(e.g. `request_scores_nfl()`).

This function is not really intended to be called directly.
Instead, it is expected that `get_scores()` will be called.
Nonetheless, this function is provided for debugging purposes (not just
for the developer, but for the user).
}
