% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff-request.R
\name{request_ff}
\alias{request_ff}
\title{Retrieve fantasy football data}
\usage{
request_ff(url, ...)
}
\arguments{
\item{url}{character. Required.}

\item{...}{dots. Not currently used.}
}
\value{
response object.
}
\description{
Retrieve fantasy football data.
}
\details{
See the `make_url*()` function for more details.s
}
