% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores-request.R
\name{request_scores}
\alias{request_scores}
\alias{request_scores_nfl}
\title{Retrieve scores}
\usage{
request_scores(league, url, ...)

request_scores_nfl(league = "nfl", ...)
}
\arguments{
\item{league}{character. Name of league for which to get scores. One of `nfl` or `nba`.
For `_[league]()` functions, set to appropriate league by default.
(e.g. For `_nfl()`, set to `nfl`.)
Note that only `nfl` is currently implemented.}

\item{url}{character. Required.}

\item{...}{dots. Not currently used.}
}
\value{
response object.
}
\description{
Retrieve scores.
}
\details{
Note that the underlying implementation of this function does not
differ accross leagues. Nonetheeless, suffixed version (i.e. `request_scores_[league]()`)
are provided in order to main consistency with the other classes of functions
(e.g. `mak_url_scores_[league]()`).
}
