% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores-get.R
\name{get_scores}
\alias{get_scores}
\alias{get_scores_nfl}
\title{Get scores}
\usage{
get_scores(league, url = NULL, resp = NULL, ...)

get_scores_nfl(league = "nfl", ...)
}
\arguments{
\item{league}{character. Name of league for which to get scores. One of `nfl` or `nba`.
For `_[league]()` functions, set to appropriate league by default.
(e.g. For `_nfl()`, set to `nfl`.)
Note that only `nfl` is currently implemented.}

\item{url, resp}{Functionally equivalent to the arguments in the `make_url_scores()` and
`request_scores()` functions. However, not required for this function.
If provided, then used and function that would be called is skipped.
(e.g. If `resp` is provided, then `make_url_scores()` and `request_scores()`
are skipped. Instead, `parse_scores()` is called directly.)}

\item{...}{dots. Passed to wrapped functions.}
}
\value{
tibble, with the following columns: TBD.
}
\description{
Get scores.
}
\details{
This functions calls `make_url_scores()`, followed by`request_scores()`,
followed by `parse_scores()`.
}
