% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{make_url_scores}
\alias{make_url_scores}
\title{Build a url}
\usage{
make_url_scores(league = c("nfl", "nba"), ..., lang = "en", region = "us",
  calendartype = "blacklist", limit = 100, year = format(Sys.Date(),
  "\%Y"), dates = NULL, seasontype = NULL, week = NULL)
}
\arguments{
\item{league}{character(s). Required for url.}

\item{...}{dots. Not currently used.}

\item{lang, region, calendartype}{character(s). Required. Reasonable defaults are provided.}

\item{limit}{numeric or character. Part of query string. Required.}

\item{year}{character. Equivalent to `dates` in the actual API.
Provided as an option because it is more intuitive.}

\item{dates}{character. May or may not be required. Should be formatted as year.}

\item{seasontype}{integer or numeric. See details.}

\item{week}{integer. Only required for NFL scores.}
}
\value{
character representing a url.
}
\description{
Build a url.
}
\details{
Same details as the `request_scores()` function...

Note that although the API requires `"sport"` as a predecesor to league, this package
abstracts this choice away from the user because `"league"` is more likely what the
user wants to specify. `"sport"` can be derived from `"league"` internally.
}
